Added CUSTOM and FILE-POSITION types to the DEFBINARY macro.
The CUSTOM type provides another way to explicitly specify the code
to be used to read and write the data. Unlike the :READER and :WRITER
field options, CUSTOM can be used as the type of array elements, as
the type of a pointer's target, or it can be generated by an EVAL
type expander.

The FILE-POSITION type simply resolves to the result of calling
CL:FILE-POSITION on the stream at that point in the stream. Nothing
is actually read or written.

Also added the READ-BINARY-TYPE and WRITE-BINARY-TYPE functions, which
allow the programmer to read data by specifying its type, including the
extended virtual types supported by the DEFBINARY macro. It does
the same thing that the EVAL type expander does, except as a pair
of stand-alone functions instead of having to be incorporated
into a struct reader.

Also finished the EXIF demo. Exif.lisp is now a working app that
can remove the EXIF metadata from JPEG files.

Still TODO at this point: Add several of the newest types to
DEFBINARY's docstring, and write a proper README.md and a
tutorial.